requirements.txt
asgiref==3.8.1
Django==5.1.5
python-decouple==3.8
sqlparse==0.5.3
xlrd==2.0.1
gunicorn==20.1.0
djongo==1.3.6
pymongo==3.11.4
django-widget-tweaks==1.4.1
pytz==2023.3
six==1.16.0
xlwt==1.3.0
